Extract copy functionality and tooltip into reusable component TextCopier
